rename: types 폴더명 type으로 수정
